+++ src/declaration/index.d.ts
@@ -17,7 +17,7 @@ export type EventMapOf<E> = E extends SVGElement
   : E extends Window
   ? WindowEventMap
   : E extends Document
-  ? DocumentEventMap
+  ? DocumentEventMap | { DOMContentLoaded: Event }
   : { [key: string]: Event };
 
 /** Gets event based on element and event type. */
@@ -405,6 +405,52 @@ export type AttributeNamesOf<E extends Element> = E extends HTMLFormElement
   ? LiAttributeNames
   : GlobalAttributeNames;
 
+/** [Link to list.](https://html.spec.whatwg.org/#attributes-3) */
+export type BooleanAttributesOf<T extends Element> = T extends HTMLIFrameElement
+  ? 'allowfullscreen'
+  : T extends HTMLScriptElement
+  ? 'async' | 'defer' | 'nomodule'
+  : T extends HTMLAudioElement
+  ? 'autoplay' | 'controls' | 'loop' | 'muted'
+  : T extends HTMLVideoElement
+  ? 'autoplay' | 'controls' | 'loop' | 'muted' | 'playsinline'
+  : T extends HTMLInputElement
+  ?
+      | 'checked'
+      | 'disabled'
+      | 'formnovalidate'
+      | 'multiple'
+      | 'readonly'
+      | 'required'
+  : T extends HTMLTrackElement
+  ? 'default'
+  : T extends HTMLButtonElement
+  ? 'disabled' | 'formnovalidate'
+  : T extends HTMLOptGroupElement
+  ? 'disabled'
+  : T extends HTMLOptionElement
+  ? 'disabled' | 'selected'
+  : T extends HTMLSelectElement
+  ? 'disabled' | 'multiple' | 'required'
+  : T extends HTMLTextAreaElement
+  ? 'disabled' | 'readonly' | 'required'
+  : T extends HTMLFieldSetElement
+  ? 'disabled'
+  : T extends HTMLLinkElement
+  ? 'disabled'
+  : T extends HTMLImageElement
+  ? 'ismap'
+  : T extends HTMLFormElement
+  ? 'novalidate'
+  : T extends HTMLDetailsElement
+  ? 'open'
+  : T extends HTMLDialogElement
+  ? 'open'
+  : T extends HTMLOListElement
+  ? 'reversed'
+  : // Common for most elements.
+    'autofocus' | 'hidden' | 'itemscope';
+
 /**
  * Select element from parent node. By default selects from **document**.
  * Safe variant of `document.querySelector` method.
@@ -575,7 +621,7 @@ export function removeEventListener<
  */
 export function toggleAttribute<E extends Element>(
   element: E | Maybe<E>,
-  name: AttributeNamesOf<E> | GlobalAttributeNames,
+  name: BooleanAttributesOf<E>,
   force?: boolean
 ): boolean;
 export function toggleAttribute<E extends Element>(