+++ test/clone_node.test.ts
@@ -8,7 +8,7 @@ describe('cloneNode', () => {
   test('cloneNode clones node with default deep value', () => {
     expect(
       querySelector('p')
-        .map(cloneNode)
+        .chain(cloneNode)
         .map((clonedP) => {
           expect((clonedP as Element).innerHTML).toMatch('');
         })
@@ -18,7 +18,7 @@ describe('cloneNode', () => {
   test('cloneNode clones node with descendants', () => {
     expect(
       querySelector('p')
-        .map((p) => cloneNode(p, true))
+        .chain((p) => cloneNode(p, true))
         .map((clonedP) => {
           expect((clonedP as Element).innerHTML).toMatch('<a></a>');
         })